refactor: 공통으로 쓰는 models, hooks, constants를 Common으로 이동
